Struct wl_copy::Options

pub(crate) struct Options {
    paste_once: bool,
    foreground: bool,
    clear: bool,
    primary: bool,
    regular: bool,
    trim_newline: bool,
    seat: Option<String>,
    mime_type: Option<String>,
    text: Vec<OsString>,
    verbose: usize,
}

Fields
paste_once: bool

Serve only a single paste request and then exit

This option effectively clears the clipboard after the first paste. It can be used when copying e.g. sensitive data, like passwords. Note however that certain apps may have issues pasting when this option is used, in particular XWayland clients are known to suffer from this.

foreground: bool

Stay in the foreground instead of forking

clear: bool

Clear the clipboard instead of copying

primary: bool

Use the “primary” clipboard

Copying to the “primary” clipboard requires the compositor to support the data-control protocol of version 2 or above.

regular: bool

Use the regular clipboard

Set this flag together with –primary to operate on both clipboards at once. Has no effect otherwise (since the regular clipboard is the default clipboard).

trim_newline: bool

Trim the trailing newline character before copying

This flag is only applied for text MIME types.

seat: Option<String>

Pick the seat to work with

By default wl-copy operates on all seats at once.

mime_type: Option<String>

Override the inferred MIME type for the content

text: Vec<OsString>

Text to copy

If not specified, wl-copy will use data from the standard input.

verbose: usize

Enable verbose logging